+++ TestLib1/Library.fs
@@ -1,6 +1,6 @@
-﻿namespace TestLib
+﻿namespace TestLib1
 
 module Say =
     let hello11 name =
-        printfn "He l %s" name
+        printfn "He   l %s" name