+++ TestLib2/Library.fs
@@ -2,5 +2,5 @@ namespace TestLib2
 
 module Say =
     let hello2 name = 
-        printfn "He llo %s" name
+        printfn "He llo  %s" name